requirements.txt
# Core
numpy>=1.26
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
opencv-python>=4.10
streamlit>=1.55

# Training/inference stack
ultralytics>=8.4.0

# Experiment tracking
mlflow>=2.13

# Phase 2: VLM reasoning (PaliGemma 3B NF4)
transformers>=4.40
bitsandbytes>=0.43
accelerate>=0.30

# Phase 2: MLOps & Edge
paho-mqtt>=2.0
scikit-image>=0.22

# Phase 2: VLM caption generation (offline tool)
google-generativeai>=0.5

# Production Pipeline
onnxruntime-gpu>=1.17        # ONNX Runtime (CUDA + TensorRT EP)
fastapi>=0.111               # Production HMI backend
uvicorn>=0.30                # ASGI server for FastAPI
asyncua>=1.1                 # OPC-UA server (IEC 62541)
pymodbus>=3.6                # Modbus TCP (PLC communication)

# CUDA 12.1 PyTorch (Windows target)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121; platform_system == "Windows"
torchvision==0.20.1+cu121; platform_system == "Windows"
torchaudio==2.5.1+cu121; platform_system == "Windows"

# Fallback for non-Windows environments
torch>=2.5.0; platform_system != "Windows"
torchvision>=0.20.0; platform_system != "Windows"
torchaudio>=2.5.0; platform_system != "Windows"
